+++ internal/model/item.go
@@ -2,12 +2,10 @@ package model
 
 // Item holds item configuration for a Docker image
 type Item struct {
-	Image       string   `yaml:"image,omitempty"`
-	RegCredID   string   `yaml:"reg_cred_id,omitempty"`
-	InsecureTLS bool     `yaml:"insecure_tls,omitempty"`
-	WatchRepo   bool     `yaml:"watch_repo,omitempty"`
-	IncludeTags []string `yaml:"include_tags,omitempty"`
-	ExcludeTags []string `yaml:"exclude_tags,omitempty"`
-	Timeout     int      `yaml:"timeout,omitempty"`
-	RegCred     RegCred  `json:"-"`
+	Image       string   `yaml:"image,omitempty" json:",omitempty"`
+	RegistryID  string   `yaml:"registry_id,omitempty" json:",omitempty"`
+	WatchRepo   bool     `yaml:"watch_repo,omitempty" json:",omitempty"`
+	IncludeTags []string `yaml:"include_tags,omitempty" json:",omitempty"`
+	ExcludeTags []string `yaml:"exclude_tags,omitempty" json:",omitempty"`
+	Registry    Registry `yaml:"-" json:"-"`
 }